Upgrade to dear imgui v1.88
This was surprisingly hard to do. thprac relies on a lot of legacy
things that were removed in newer version of dear imgui. thprac also
relies on a lot of modifications to dear imgui to work.

To do this upgrade I had to make a few sacrifices.

1st: the custom titlebar and client side decorations are gone and
replaced with the default Windows titlebar. This is due to the fact that
the way the minimize button works relies on a thprac specific
modification to ImGui::Begin.

2nd: keyboard and controller controls are broken. The way thprac
achieves keyboard controls to begin with is one of the major things that
relies on legacy features and thprac specific modifications to dear
imgui itself.

Perhaps upgrading to a new version of dear imgui isn't worth it now, and
will only become worth it once dear imgui v2.00 rolls around

I have also deleted ImPlot, and all data collection and plotting
features, as well as the L menu in th13. I think those should be their
own separate program anyways